Fixed isseues
