+++ src/views/Weather-Components/Constant.js
@@ -13,14 +13,14 @@ export const Cities = [
   { title: "Attock", value: "attock" },
 ];
 
-export const popularCities = [
+export const POPULAR_CITIES = [
   { city: "Lahore", province: "Punjab" },
   { city: "Multan", province: "Punjab" },
   { city: "Kirachi", province: "Sindh" },
   { city: "Quetta", province: "Balochistan" },
 ];
 
-export const Buttons = [
+export const DETAIL_TABLS = [
   { name: "Today" },
   { name: "Hourly" },
   { name: "Daily" },
